reorganize settings crate a bit
